+++ bin/install_cosign.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 
-COSIGN_RELEASE=v1.13.1
+COSIGN_RELEASE=v1.14.0
 INSTALL_DIR=$HOME/.cosign
 
 RUNNER_OS=$(uname)